rhetoric/scripts/ai-counsel/requirements.txt
# ai-counsel/requirements.txt
# Core dependencies
pydantic>=2.5.0
pyyaml>=6.0.1

# HTTP client and retry logic for HTTP adapters
httpx>=0.27.0
tenacity>=8.2.0

# Test dependencies
pytest>=7.0
pytest-asyncio>=0.21
respx>=0.20  # For mocking httpx

# Enhanced convergence detection backends
# TF-IDF similarity backend (good performance, moderate accuracy)
scikit-learn>=1.3.0

# Neural semantic similarity backend (best performance, highest accuracy)
# Provides most accurate convergence detection and vote grouping
sentence-transformers>=2.2.0
